Assignment_1/code/mysimbdp-dataingest/requirements.txt
confluent-kafka
cassandra-driver
argparse
pandas